Further expand node weight support
Should fully support skinned mesh importing

Changes to how we get the parent node for markers. Now reads the parent bone the object has instead of checking the vertex group assigned in index 0

Add a way to define region and permutation names in the materials tab. Will probably be changed to avoid an issue with unique material names.

Change how we get marker location to be a bit more accurate.  Still needs work.