More physics work.
Some form of point to point support. Not complete but it will do for now.

Stop script from attempting to read the parent of a null object in physics constraints.

Stop script from attempting to read a material out of list range

Use a boolean in JMS object properties instead of a affix symbol for bounding sphere since we don't actually know how this originally worked.