Complete hidden geo option and a few bug fixes
This commit introduces the hidden geo behavior from Blitzkrieg. If an object is hidden either through hide_set or hide_viewport then it will not be included on export. There is also an option in the export menu to ignore whether or not they are hidden and export everything in the scene. Still need to handle the exclude property and hide_set for collections. This will come later.

Finished vert code so that it would work properly for armatures. Objects before weren't getting weights assigned if it was an armature.

Reworked the UI a bit more. There is now a global game version option in scene properties that handles how the UI is displayed for all objects.

Fixed matrix not being defined for convex shapes.

Removed some not used variables and merged some imports.

Changed to value added to instance marker ID.

Made sure <None> material is only added when needed.

Clean up some of the functions so that return isn't spammed everywhere. Should be a bit more clear maybe.