Make sure we aren't reading invalid memory
Clear to_mesh() mesh data once we are done with it. The data wasn't cleared when the script finished and caused subsequent runs to potentially be unstable

Fix custom scale not applying to vertex locations

Cancel script if an error is raised

Normalize the results from the normal calculations

Only get index when checking for a material in object loop. Don't need to get the material slot.