Add options to simplify FP importing
Revert return changes in export_jms and export_jma. I made a mistake in why an error was happening.

Add menu options to make it easier to import FP anims. Will start working for docs at some point @csauve offered to attempt to make a better alternative to the curent UI setup for it. Either way it seems to work ok as it is atm to commit.