Lightmapper prep changes and new sort method
Changes to lightmapper code. Still needs a few things but it's getting there.

Added a new sort method that finds an object and uses the vertex group list to sort the node list. This is so that the exporter can be used with the Halo 2 Vista render model tool known as DAEconverter.