Fix for WRL importer errors
If an object already exists in the scene then make sure we set it to object mode so that we can do to_mesh() properly.

Make sure we append the material if either the object doesn't exist but the material does or the object does exist but the material isn't assigned.