Fix generated images writing whitespace
Generated images in Blender would cause the exporter to write whitespace for the texture path since it isn't an actual file on disk. We now switch to getting the name of the image node if the image source isn't a file.

Also break after finding the first valid image node. No reason to continue.