Undo tool region rename
Rename "__unnamed" back to "unnamed" on import. Tool adds underscores to this region for whatever reason.

Change how WRL importing gets the mesh and object. To_mesh was never happening if there was a single face or edge because there would need to be at least two loops for it to change the variable from none to an object.