Improve animation import times
No reason to call update(). Was just slowing things down.

Set the JMS rest position import in JMA to use auto for game version. Not sure while this was set to Halo 2 specifically but maybe we'll find out.

Enable object names for imported WRL objects and set origin to geometry.