Improve skylight generator a bit.
Weaken the color of the skylights further away from the sun point.

Mess with rotation value a bit so that sunlights don't overlap

Generate a single white skylight for our shadows.

Allow users to define the sun in their scene through yaw and pitch. Will need something to represent this somehow.