ASS and JMS fixes
Make sure the instance list starts with none to represent the scene root in ASS export

use the right name for light exporting

Move default names for material definition to the start of the object loop

Define the armature if object uses an armature as a parent

Match Halo 3 ASS file when it comes to writing weights

Fix memory leak that caused the exporter to use more ram than it should. Now how did that happen?