Toolset cleanup
Add support for QUA exporting. Super early and probably not using the options like they are supposed to. Need to look into it more.

Remove encoding arg. Ancient history now lost to the sands of time.

Remove some useless imports and merge a few.

Remove useless args in some functions along with unused code.

Use the passed context instead of getting it from bpy. If there a difference? I have no idea but it's already there so why not use it instead.

Remove duplicate code in get_dimensions

Remove hack fix for scale issues in brute.ASS. This wasn't meant to be in a public build and just breaks more things.

Readd support for custom scale on export. Needs more work before it's useable for ASS but it should be good in JMS.

Replace commas with periods in vectors and quaternions to get around locale issue in Bonobo files.

Fix cleaning and normalize weights option not doing anything.