Fixes for new features
Don't break when the base_map is found. Check if it's valid first.

Make sure we've found a valid bitmap path before trying to use it

Process H1 material name correctly

Only generate a skeleton for an ASS on import if the ASS file is Halo 3 and it's not in a structure directory

Set the game title on import in JMS and ASS

Add ODST paths. Does nothing right now

Replace try/catch stuff with something better