More lightmap fixes
If an exact shader group is not found and a fallback isn't found either then fallback to just generating a simple shader (a bdsf with the first bitmap it can find.)

Add blacklist to shader group search. Maybe needs an option on the Blender side?

Misc cleanup for Halo 2 shader resource blend

Add a denoise step to lightmaps.