Fix region issues in ASS imports
Make sure we only create regions per mesh

Put the entire string into the region without any changes. The code was more meant for JMS and there's no real reason for us to pull it apart. As it is it was changing some contents for strings since it doesn't expect more than 4 args and material properties were showing up